requirements.txt

networkx>=2.3,<3
PyYAML>=3.11
snakeviz
